refactor(core): migrate User ID to long and refine Java management UI
- Change User ID data type from int to long in UserModel and repositories for Telegram compatibility
- Refine inner logic and service handling in JavaArgumentManagementViewModel and JavaManagerService
- Update WindowService and repository calls to match updated data models

+++ TelegramBotMinecraft.Core/Models/UsersModel.cs
@@ -4,9 +4,9 @@
     {
         public string Name { get; set; }
 
-        public int Id { get; set; }
+        public long Id { get; set; }
 
-        public User(string name, int id)
+        public User(string name, long id)
         {
             Name = name;
             Id = id;
